[PYOPENMS,FIX] fixed streampos->long conversion

+++ pyOpenMS/addons/streampos.pyx
@@ -1,4 +1,8 @@
 
 
     def __long__(self):
-        return <long>(deref(self.inst.get()))
+        # <long> calls c++ conversion streampos -> c++ long value
+        # long(..) converts this value to a Python long.
+        # value. else this method may return an int or long, depending on the magnitude of
+        # the value !
+        return long(<long>(deref(self.inst.get())))
